+++ packages/core/src/utils/transformers/media/transformMedia.ts
@@ -9,7 +9,8 @@ import {
 import { Asset, AssetFile } from 'contentful';
 import { getOptimizedBackgroundImageAsset } from './getOptimizedBackgroundImageAsset';
 import { getOptimizedImageAsset } from './getOptimizedImageAsset';
-import { getBoundValue } from '../getBoundValue';
+import { getBoundValue } from '@/utils/transformers/getBoundValue';
+import { isDeepPath, lastPathNamedSegmentEq } from '@/utils';
 import { ValidFormats } from './mediaUtils';
 
 export const transformMedia = (
@@ -20,6 +21,12 @@ export const transformMedia = (
   path: string,
 ) => {
   let value: BoundComponentPropertyTypes;
+
+  // If it is not a deep path and not pointing to the file of the asset,
+  // it is just pointing to a normal field and therefore we just resolve the value as normal field
+  if (!isDeepPath(path) && !lastPathNamedSegmentEq(path, 'file')) {
+    return getBoundValue(asset, path);
+  }
 
   //TODO: this will be better served by injectable type transformers instead of if statement
   if (variableName === 'cfImageAsset') {
@@ -83,6 +90,5 @@ export const transformMedia = (
     return;
   }
 
-  // return getBoundValue(asset, entityStore, binding, path);
-  return getBoundValue(asset, path);
+  return asset.fields.file?.url;
 };