+++ packages/experience-builder-sdk/src/blocks/editor/VisualEditorRoot.tsx
@@ -2,20 +2,23 @@ import React, { Suspense } from 'react';
 import { EntityStore, VisualEditorMode } from '@contentful/experiences-core';
 import { ErrorBoundary } from '../../components/ErrorBoundary';
 import { useInitializeVisualEditor } from '../../hooks/useInitializeVisualEditor';
+import type { Experience } from '@contentful/experiences-core/types';
 
 const VisualEditorLoader = React.lazy(() => import('./VisualEditorLoader'));
 
 type VisualEditorRootProps = {
   visualEditorMode: VisualEditorMode;
-  initialEntities: EntityStore['entities'];
+  experience?: Experience<EntityStore>;
   initialLocale: string;
 };
 
 export const VisualEditorRoot: React.FC<VisualEditorRootProps> = ({
   visualEditorMode,
-  initialEntities,
+  experience,
   initialLocale,
 }) => {
+  const initialEntities = experience?.entityStore?.entities || [];
+
   useInitializeVisualEditor({
     initialLocale,
     initialEntities,
@@ -24,7 +27,7 @@ export const VisualEditorRoot: React.FC<VisualEditorRootProps> = ({
   return (
     <ErrorBoundary>
       <Suspense fallback={<div>Loading...</div>}>
-        <VisualEditorLoader visualEditorMode={visualEditorMode} />
+        <VisualEditorLoader experience={experience} visualEditorMode={visualEditorMode} />
       </Suspense>
     </ErrorBoundary>
   );