+++ packages/test-app/src/eb-config.ts
@@ -1,5 +1,7 @@
 import { defineComponents } from '@contentful/experiences-sdk-react';
 import ComponentWithChildren from './components/ComponentWithChildren';
+import { LinkComponent } from './components/LinkComponent';
+import { CustomImageComponent } from './components/CustomImageComponent';
 
 defineComponents([
   {
@@ -20,6 +22,55 @@ defineComponents([
     },
     options: {
       wrapComponent: false,
+    },
+  },
+  {
+    component: CustomImageComponent,
+    definition: {
+      id: 'custom-image',
+      name: 'Custom Image',
+      category: 'Custom Components',
+      variables: {
+        src: {
+          displayName: 'SRC',
+          type: 'Media',
+        },
+      },
+    },
+  },
+  {
+    component: LinkComponent,
+    definition: {
+      id: 'link-component-custom',
+      name: 'Link Component with custom link',
+      hyperlinkPattern: 'https://test.com/{locale}',
+      variables: {
+        href: {
+          displayName: 'Href',
+          type: 'Hyperlink',
+        },
+        text: {
+          displayName: 'Text',
+          type: 'Text',
+        },
+      },
+    },
+  },
+  {
+    component: LinkComponent,
+    definition: {
+      id: 'link-component',
+      name: 'Link Component',
+      variables: {
+        href: {
+          displayName: 'Href',
+          type: 'Hyperlink',
+        },
+        text: {
+          displayName: 'Text',
+          type: 'Text',
+        },
+      },
     },
   },
 ]);