+++ packages/visual-editor/src/components/Dropzone/Hitboxes.tsx
@@ -1,8 +1,8 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useMemo } from 'react';
 import styles from './styles.module.css';
 import { useTreeStore } from '@/store/tree';
 import { getItemDepthFromNode } from '@/utils/getItem';
-import { CTFL_ZONE_ID, HitboxDirection, ROOT_ID } from '@/types/constants';
+import { CTFL_DRAGGING_ELEMENT, CTFL_ZONE_ID, HitboxDirection, ROOT_ID } from '@/types/constants';
 import { useZoneStore } from '@/store/zone';
 import { useDraggedItemStore } from '@/store/draggedItem';
 import { createPortal } from 'react-dom';
@@ -11,10 +11,10 @@ import { getHitboxStyles } from '@/utils/getHitboxStyles';
 interface Props {
   parentZoneId: string;
   zoneId: string;
-  enableRootHitboxes: boolean;
+  isEmptyZone: boolean;
 }
 
-const Hitboxes: React.FC<Props> = ({ zoneId, parentZoneId, enableRootHitboxes }) => {
+const Hitboxes: React.FC<Props> = ({ zoneId, parentZoneId, isEmptyZone }) => {
   const tree = useTreeStore((state) => state.tree);
   const isDraggingOnCanvas = useDraggedItemStore((state) => state.isDraggingOnCanvas);
   const scrollY = useDraggedItemStore((state) => state.scrollY);
@@ -22,39 +22,42 @@ const Hitboxes: React.FC<Props> = ({ zoneId, parentZoneId, enableRootHitboxes })
     () => getItemDepthFromNode({ id: parentZoneId }, tree.root),
     [tree, parentZoneId],
   );
-  const [fetchDomRect, setFetchDomRect] = useState(Date.now());
-
-  useEffect(() => {
-    /**
-     * A bit hacky but we need to wait a very small amount
-     * of time to fetch the dom getBoundingClientRect once a
-     * drag starts because we need pre-drag styles like padding
-     * applied before we calculate positions of hitboxes
-     */
-    setTimeout(() => {
-      setFetchDomRect(Date.now());
-    }, 50);
-  }, [isDraggingOnCanvas]);
+  const zones = useZoneStore((state) => state.zones);
+  const hoveringZone = useZoneStore((state) => state.hoveringZone);
+  const isHoveringZone = hoveringZone === zoneId;
 
   const hitboxContainer = useMemo(() => {
     return document.querySelector('[data-ctfl-hitboxes]');
   }, []);
 
   const domRect = useMemo(() => {
+    if (!isDraggingOnCanvas) return;
     return document.querySelector(`[${CTFL_ZONE_ID}="${zoneId}"]`)?.getBoundingClientRect();
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [zoneId, fetchDomRect]);
+  }, [zoneId, isDraggingOnCanvas]);
 
-  const zones = useZoneStore((state) => state.zones);
+  // Use the size of the cloned dragging element to offset the position of the hitboxes
+  // So that when dragging causes a dropzone to expand, the hitboxes will be in the correct position
+  const offsetRect = useMemo(() => {
+    if (!isDraggingOnCanvas || isEmptyZone || !isHoveringZone) return;
+    return document.querySelector(`[${CTFL_DRAGGING_ELEMENT}]`)?.getBoundingClientRect();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [isEmptyZone, isHoveringZone, isDraggingOnCanvas]);
 
   const zoneDirection = zones[parentZoneId]?.direction || 'vertical';
   const isVertical = zoneDirection === 'vertical';
   const isRoot = parentZoneId === ROOT_ID;
-  const showRootHitboxes = isRoot && enableRootHitboxes;
 
   const getStyles = useCallback(
-    (direction: HitboxDirection) => getHitboxStyles({ direction, zoneDepth, domRect, scrollY }),
-    [zoneDepth, domRect, scrollY],
+    (direction: HitboxDirection) =>
+      getHitboxStyles({
+        direction,
+        zoneDepth,
+        domRect,
+        scrollY,
+        offsetRect,
+      }),
+    [zoneDepth, domRect, scrollY, offsetRect],
   );
 
   const ActiveHitboxes = (
@@ -66,14 +69,13 @@ const Hitboxes: React.FC<Props> = ({ zoneId, parentZoneId, enableRootHitboxes })
           isVertical ? HitboxDirection.SELF_VERTICAL : HitboxDirection.SELF_HORIZONTAL,
         )}
       />
-      {showRootHitboxes && (
+      {isRoot ? (
         <div
           data-ctfl-zone-id={parentZoneId}
           className={styles.hitbox}
           style={getStyles(HitboxDirection.BOTTOM)}
         />
-      )}
-      {!isRoot && (
+      ) : (
         <>
           <div
             data-ctfl-zone-id={parentZoneId}